Move tests expectations to module
Easier tracking of log message requirements for requests to the API
